firestarter/requirements.txt
# Core Frameworks
langchain>=0.1.0
langchain-ollama>=0.1.0
langgraph>=0.0.40
pyautogen>=0.2.0
llama-index>=0.10.0

# Models & Infrastructure
ollama>=0.1.0
playwright>=1.40.0
google-search-results>=2.4.2
ddgs>=0.1.0  # DuckDuckGo search (free fallback for web search)
# Note: Also supports duckduckgo-search package for backward compatibility

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter for conversation storage
pgvector>=0.2.0  # PostgreSQL vector extension for semantic search
redis>=5.0.0  # Redis for short-term memory buffer

# Tools & Utilities
pydantic>=2.0.0
python-nmap>=0.7.1
pymetasploit3>=1.0.0
shodan>=1.31.0
virustotal-api>=1.1.11

# Data Processing
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
lxml[html_clean]>=4.9.0
requests>=2.31.0
jinja2>=3.1.0
bs4>=0.0.1  # BeautifulSoup4 alias

# Additional dependencies
python-dotenv>=1.0.0
rich>=13.0.0
typing-extensions>=4.8.0

# Fuzzy matching
rapidfuzz>=3.0.0
python-Levenshtein>=0.21.0
